[A11y] Don't discard aria-hidden for bad focus until after a delay
Angular Material and other UIs sometimes move the focus a clock tick
after the aria-hidden is set. This results in Chromium's repair code
being activated, which logs a console error and discards the
aria-hidden. Since this is a somewhat common practice, a decision has
been made to make the error message slightly more forgiving.
We allow one serialization to occur with the illegal focus, and
the second serialization then requires focus to be moved out of
the aria-hidden subtree.

Example to play with: https://codebin.googleplex.com/#/jhsbbn2pjk3

tl;dr. Give the web page a moment to move the focus to somewhere appropriate (not inside aria-hidden).

Fixed: 392121909
Change-Id: Ie6791db6328b3af73670159436284d1de848798a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6215572
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1416505}
NOKEYCHECK=True
GitOrigin-RevId: c39c575d5b94643ed8115c501893cc35c1a91db8