Null check `relocatable_end_of_next_paragraph` before updating it
This CL checks if `relocatable_end_of_next_paragraph` has a valid
position before updating it for adjusting a list item indentation
at FormatSelection(). It can be null when FormatSelection() is
executed at the end of paragraph as it tries to get the next
position of the current paragraph.

Bug: 393411220
Change-Id: Id4f1868d81f4f9cf509e1af2691cd887672ccab9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6234864
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Commit-Queue: Julie Jeongeun Kim <jkim@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1416610}
NOKEYCHECK=True
GitOrigin-RevId: 74bd1b42d125e9aaf1435f130bcde93a5c61afb9