Relax option OwnerSelectElement checking
In commit 233b8b3 I made OptionList capable of setting
OwnerSelectElement when it hadn't been set yet in case not every option
inside the select element had InsertedInto called on it yet. I did this
only during a document fragment parsing fast path because I thought that
was the only time this was possible, but based on crash reports, there
are more cases. This patch makes the OptionList OwnerSelectElement
assignment happen regardless of whether we are in the document fragment
parsing fast path or not.

Fixed: 393579139
Change-Id: I775fd43a52b9bb7f5379166bd75f953dfd95c2e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6231131
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1416831}
NOKEYCHECK=True
GitOrigin-RevId: 44c9a3bac1eb29b78b55d247b64727f096053b36