Move Font to Oilpan.
Currently, Font is a DISALLOW_NEW() object that is part of other objects
(in particular, the StyleFontData group in ComputedStyle); however, we
need to keep pointers to it from e.g. CSSLengthConversionData, which is
not allowed and could cause use-after-free if nothing else keeps the
Font object alive. (There's a GC_PLUGIN_IGNORE for this, which we want
to remove in this CL.)

However, StyleFontData (the “font” group in ComputedStyle) contains
_only_ Font; nothing else. So we can simply combine the two classes;
we put Font on Oilpan, and replace Member<StyleFontData> with
Member<Font>. This means we don't really get more Oilpan objects
floating around, which means performance is generally unaffected.
(We could consider moving the Member<Font> into the normal list of
inherited objects.)

The alternative we originally considered was moving FontDescription
to Oilpan, but that was a _lot_ more involved (e.g., a lot of code
assumes that it is default-constructible with a sane value), create
more Oilpan objects, and would require inlining Font onto
CSSLengthConversionData etc..

This patch is for correctness and not for performance, but for some
reason I don't fully understand, parsing becomes ~1% faster in the
style perftest.

Bug: 389707046
Fixed: 391114794
Change-Id: I0d6262e751aacfb4ab6706ec7932300a49175416
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6234313
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1416946}
NOKEYCHECK=True
GitOrigin-RevId: a479c24d1ea0093435ba8bef5a365822ba6c6c6e