[css-highlight-api] Check Node::isConnected in ValidateHighlightMarkers
Check if the Highlight's AbstractRange start and end nodes are connected
before creating an EphemeralRange to create a new marker in
HighlightRegistry::ValidateHighlightMarkers().

There were cases when a Highlight's range was anchored in a node under a
shadow tree and the node was removed, then this range's start and end
nodes were disconnected, causing the creation of an EphemeralRange with
disconnected start and end positions in ValidateHighlightMarkers(). See
the bug for more details.

Add a crash test for this case.

Bug: 394634479
Change-Id: I9e487dbbf3f1f4ce5049f84b7616560d892de9e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6236739
Reviewed-by: Sanket Joshi <sajos@microsoft.com>
Commit-Queue: Fernando Fiori <ffiori@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1417405}
NOKEYCHECK=True
GitOrigin-RevId: 7d64db0b77284a63ea6584949ad6a0ac8a643320