[AriaNotify] Update names of AriaNotificationPriority values
This CL updates the values of the "priority" property on the ARIA
Notify API from "none" to "normal" and "important" to "high". The
default value of this property is now "normal". This comes from a
decision from the ARIA Working Group:
https://github.com/w3c/aria/issues/2333

ARIA Notify explainer:
https://github.com/MicrosoftEdge/MSEdgeExplainers/blob/main/Accessibility/AriaNotify/explainer.md

Bug: 326277796
Change-Id: I388ed11c0750aff9e4f00a2973f5b769229a9bdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6237614
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Commit-Queue: Evelynn Kaplan <evelynn.kaplan@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1417513}
NOKEYCHECK=True
GitOrigin-RevId: 10def2a1b12e2d5a92d4a0d8003c121f9f2429a7