[BCIV] Refactor BrowserControlsOffsetTagsInfo
BrowserControlsOffsetTagsInfo consists of a bunch of things, and is
passed from java to DFHA and renderer, but neither DFHA or the renderer
needs everything in BrowserControlsOffsetTagsInfo.

Instead of always passing a BrowserControlsOffsetTagsInfo to clients,
this CL creates a BrowserControlsOffsetTagDefinitions (sent to DFHA) and
BrowserControlsOffsetTagModifications (sent to renderer), which consists
of only the necessary information needed by the clients.

More specifically, the DFHA only needs the OffsetTags and
OffsetTagConstraints, while the renderer only needs the OffsetTags and
the top/bottom controls' additional heights.

Bug: 393608180
Change-Id: I6141506659e2fb5efb33df0ca3045497ec141fc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6217648
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Owners-Override: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Peilin Wang <peilinwang@google.com>
Cr-Commit-Position: refs/heads/main@{#1418012}
NOKEYCHECK=True
GitOrigin-RevId: 1187e4e80bbad755ac5a2234409153dab988bb54