Ensure that attr-taint carries through if()
This is not speced yet, but the logic is following:

- if there is attr-tainted value in the if() condition, either in custom
property or query specified value, that doesn't affect the return value,
so returned value would not be tainted just because if that.

- if the picked declaration value is attr()-tainted, the return value of
if() function becomes tainted.

- unused declaration value does not affect the attr()-tainting of the
returned value.

This CL also ensures that attr()-tainting does not affect comparison in
the if() style condition.

Bug: 393698480
Change-Id: Ia3c517bb4c89de1a81ff17c40b3b7f0c850fc3fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6239675
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Munira Tursunova <moonira@google.com>
Cr-Commit-Position: refs/heads/main@{#1418112}
NOKEYCHECK=True
GitOrigin-RevId: b88b05d570036fea64e1c639dbb0203f293f07e2