Migrate media/ to start-from-one ref counting
Add REQUIRE_FOR_REFCOUNTED_TYPE to all subclases of
base::RefCountedThreadSafe in media/ and use base::MakeRefCounted
instead of `new` for object creation where needed.

In this CL, the remaining classes in media/base/ and media/audio/ are
migrated. The rest of the changes will be handled in a follow-up CL.

See
https://source.chromium.org/chromium/chromium/src/+/main:base/memory/ref_counted.h;l=309;drc=082e8ce5d3bff64458678297dffa15ea5cfc8611
for the motivations to use start-from-one ref count.

Bug: 40608481
Change-Id: I528188f3314d3eaadfdd219fa65f75bff63a11a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6198498
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Kenneth MacKay <kmackay@chromium.org>
Reviewed-by: Pilar Molina Lopez <pmolinalopez@chromium.org>
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Reviewed-by: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418202}
NOKEYCHECK=True
GitOrigin-RevId: 3252b35ef84cf52e4ddd506ebb5ceb81566a58de