[carousel] Focus after activating scroll marker is its scroll target
As per https://drafts.csswg.org/css-overflow-5/#scroll-target-focus,
when a scroll marker is activated, the next tabindex-ordered focus
navigation will focus the scroll target if it is focusable, otherwise,
it will find the next focusable element from the scroll target as though
it were focused.

This CL uses sequential navigation starting point to remember where to
start looking after scroll marker activation.

so it's left for the future CLs, and for now it will focus scroller.

Note: it's not clear how to implement it for ::column::scroll-marker,
Bug: 378698659
Change-Id: I1ff5ccdc9526f09a36faa387e8ad2f9a1dc26a54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6097493
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Daniil Sakhapov <sakhapov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418394}
NOKEYCHECK=True
GitOrigin-RevId: 64848be4dc29f522395f476df047ef2bc323fb35