Fix dialog crashers related to `<dialog open>`
When the `open` attribute is added to a dialog to force-open it, the
close watcher was previously not getting constructed. Several related
cases then could crash, due to the lack of a close watcher. This CL
constructs the close watcher, and adds the dialog to the open dialogs
list, when the `open` attribute is manually added.

Fixed: 393420204, 393883102, 393879748, 393834331
Bug: 376516550
Change-Id: Id80ac36a3b9295b071fc56c251f6f3706573bc27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6245064
Reviewed-by: Traian Captan <tcaptan@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Commit-Queue: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418682}
NOKEYCHECK=True
GitOrigin-RevId: 4ba99d6ab6593ef89703a1e69a8b965502edb96a