Fix compositor sync after pause + set time + play
https://chromium-review.googlesource.com/c/chromium/src/+/4605129
introduced a regression where a restart would preserve start time
even in cases where a new start time was required.

This patch cleans up the logic for determining if a restart is
required, and if a fresh start time is needed when restarting. If
parameters to compositing state remain unchanged and simply waiting on
a start time, we don't need to cancel or restart and can simply keep
the animation in a pending state until the start time is received. If
keyframes where changed, but timing characteristics remain the same,
we should preserve start time while restarting. This is the issue the
aforementioned patch was addressing.

The terms soft_change and hard_change were not clearly defined in
the code. This terminology has been replaced in order to make the
process clear.

Bug: 374374231
Change-Id: Ida5a9dc2c53a0f46bf33a2064c29989ec2c62d2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6234613
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418806}
NOKEYCHECK=True
GitOrigin-RevId: 24e2565b11fa1ddbdf0f10d069bb7f6c08aba3fa