[DC] Fix/Improve Digital Credentials permission policy wpt
Web platforms tests for Digital Credentials Permission Policy has the
following issues:

1- The test were very fragile. Tests were passing an incomplete request
to the API, which results in the request being rejected before even
checking the permission policy. The test were relying on the type of
error that is thrown to decide the permission policy behaviour which was
basically masks the real issues.

2- Some tests assumed that the allow="" attribute can override the
header policy which isn't correct. [*]

This CL addresses both issues by always sending a complete request to
the API, and updates the tests to reflect the latest standards.

When sending a complete request to the API, the API is expected to
forward the request to the underlying platform and request user UI. In
order to avoid test hang, one option would be to introduce a command
line flag that's enabled by default in tests, which preempts the call to
the platform APIs and responds immediately with success. This allows
testing sending proper requests to the browser without test hang.

[*]
https://www.w3.org/TR/permissions-policy/#algo-define-inherited-policy-in-container

Bug: 370982995
Change-Id: Ie2f4b94a2788691e0f62458eba41b4b1f6080c74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6239109
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418815}
NOKEYCHECK=True
GitOrigin-RevId: 21bdc744f2637dbc7c78513ec39a860c7525a5ad