compute pressure: Use associated interfaces in mojom interfaces
As described #353045204 and #371060464, there have been issues in
synchronizing messages between Blink and //content using two
interfaces blink::mojom::WebPressureManager and
device::mojom::PressureClient.

This patch addresses these issues by using associated interfaces.

blink::mojom::WebPressureManager pipe will be used as master
interface and device::mojom::PressureClient will be associated with it.

The association is done by passing the pending_associated_remote of
the clients in WebPressureManager::AddClient() function.

Similar interface changes are made also between //content and
//services, to keep the same interfaces architecture.

The unittest PressureServiceForFrameTest.AddClientTwice was redesigned
because, with associated interfaces it is impossible to add twice the
same client, due to the BindNewEndpointAndPassRemote() function.
Instead the test is trying to add a new client but with the same
pressure source.

Another related issue mentioned in #371060464 is also fixed here.

Theoretically, there can be only one WebPressureManager interface, but
in #371060464, it was shown that there can be a race when disconnecting
the last client, which will also disconnect the WebPressureManager
interface and adding a new client that will connect
a WebPressureManager interface. To solve this issue, it was decided to
keep blink::mojom::WebPressureManager alive through the whole
ExecutionContext lifetime.

Adding PressureServiceBase as WebContentPressureManagerProxy observer
is not happening when the WebPressureManager interface is bound.
Unit tests were updated accordingly.

Bug: 353045204, 371060464
Change-Id: I1b048d9d387b85cb699cae15dc18b3d2efb60e0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6084497
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Commit-Queue: Arnaud Mandy <arnaud.mandy@intel.com>
Cr-Commit-Position: refs/heads/main@{#1419114}
NOKEYCHECK=True
GitOrigin-RevId: 44dd3f8432b0dbc2186e9870400d47288dc30940