[carousel] Handle out-of-flow positioned scrollers.
If a scroller is out-of-flow, but its scroll marker group is not, they
will not be laid out in sequence. The out-of-flow scroller will
typically be laid out way after the scroll marker group. Since the
scroll marker group is populated as part of scroller layout, we need to
make sure that we only mark for layout within the subtree established by
the scroll marker group. Marking on the outside would re-mark elements
for layout, after we're done laying them out.

Also remove the early-bail for `after` scroll marker groups in
HandleScrollMarkerGroup(), since we cannot assume that such groups
actually come after in layout.

Bug: 395798192
Change-Id: I4c19afe7b3f217ff83eeebd33e1ec92d88aaee5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6253685
Reviewed-by: David Grogan <dgrogan@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1419158}
NOKEYCHECK=True
GitOrigin-RevId: d103081b49f10b92513757aa7f39ab6a40540d4c