Service Worker Client.url needs to be creation URL
The creation URL of a document represents the location of the
associated resource and this can be different from the document's
current URL if script uses the history.pushState API (see
https://html.spec.whatwg.org/multipage/webappapis.html#concept-environment-creation-url).
The creation URL should not change when the document's current URL is
modified by history.pushState. Currently the Service Worker Client
gets its url property value from RenderFrameHost's GetLastCommittedUrl
which is modified by history.pushState and so is failing the web
platform test service-workers/service-worker/clients-matchall-client-types.https
which validates this.

The ServiceWorkerClient::url() is close to the creation URL in that it
does not change when history.pushState is used. However, it has had
its URL fragment stripped. Some parts of the service worker code
require the fragment to be stripped and so this change adds a
ServiceWorkerClient::creation_url() that is the version of the URL
that does not have the fragment removed. And this creation_url is used
for the url property of the JS ServiceWorkerClient object.

Bug: 41337436
Change-Id: I3d2a50e80f0ea433f6eeb86235edf5c448729fad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6078347
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Dave Risney <david.risney@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1419436}
NOKEYCHECK=True
GitOrigin-RevId: ec3e438d82163760477dbfed5abdfe407daa4480