Check for preloaded response candidate earlier
This CL moves the check to see if there is a preloaded response
candidate in the `preloads_` list earlier. Since the PreloadKey depends
on the URL, and ModifyRequestForCSP() may update the URL, this is also
moved earlier.
This is in preparation to enable skipping irrelevant checks when we
already know we have a preloaded response candidate.

Bug: 348442535
Change-Id: Idee5df0b7410051bfcfd32aa359f07690d151b75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6233448
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Nidhi Jaju <nidhijaju@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1419736}
NOKEYCHECK=True
GitOrigin-RevId: 1ad17c95b7ce2d376c2c0ac55c13a0ccf8da9033