Disable cache-validator check for SVG images in LayoutImage
Now when we only read/update the image's natural dimensions as a result
of an image-changed notification we won't run into the issue of reading
a stale size during validation. At least not differently than for raster
images[1]. The only code branch where it applies doesn't actually have
an effect on the returned size.

Disable the mentioned code branch by adding a flag[2]. This also
disables the undoing of any DPR, since that is already performed where
the object-fit value is applied, so this would in fact mean that we're
"double-unapplying". This also mean that we treat all image types the
same way w.r.t to this.

[1] We'll still re-read the image dimensions if extrinsic state that
    affect the image dimensions change - such as the zoom factor or
    image-orientation.
[2] LayoutImageRevalidationCheckForSvgImages

Test: http/tests/svg/svgimage-revalidation.html
Bug: 40537999
Change-Id: I797b4feeba6ca78446577390d8762b2aead0e713
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6186004
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1419827}
NOKEYCHECK=True
GitOrigin-RevId: d97f345d01ad618bf265174be6fcec8874d6c2c0