[functions] Store PseudoId instead of StyleResolverState::ElementType
To implement @container support in custom functions (specifically
to implement correct container selection), it will be necessary
to expose the current PseudoId on StyleResolverState.

This CL adds PseudoId to StyleResolverState, but at the same time
removes the ElementType enum (and related member). This enum is no
longer needed, as we can determine all we need to know from the PseudoId
and ElementResolveContext.

This is a minor refactor to aid @function development; there should be
no behavior change, with or without the CSSNestedPseudoElements flag
enabled.

Bug: 325504770
Change-Id: I779a33c77d25b71317c5073c1fda3f8be26b72eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6257753
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Daniil Sakhapov <sakhapov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1419847}
NOKEYCHECK=True
GitOrigin-RevId: 3825dfe1f55ff4ffeccfdeb53f156fac1d521c1c