Update collapsible placeholder computation
This CL resolves a crash bug which occurs due to incorrect calculation
as currently we calculate the collapsible placeholder before the call to
|PrepareWhitespaceAtPositionForSplit| method which actually replaces
"\n" with a nbsp and which is why at the time of removal of collapsible
placeholder, it's not there anymore so the DCHECK fails.
Now on, the collapsible placeholder calculation would be done after all
adjustments done on insertion position and just before the node
insertion starts. This makes sure that we check for collapsible
placeholder at correct |insertion_pos|.
The web_test in this CL should not crash.

Bug: 395028747
Change-Id: Ibba1b868da27a0891a5ea8132295acdfacaed641
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6248618
Reviewed-by: Sambamurthy Bandaru <sambamurthy.bandaru@microsoft.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Siye Liu <siliu@microsoft.com>
Commit-Queue: Utkarsh Pathak <utpathak@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1419895}
NOKEYCHECK=True
GitOrigin-RevId: fb6a63a31f6fdbee0d483743b749f43e3f6aa6a6