Report sizes of string resources to V8 for heap snapshots
V8 recently added a mechanism allowing the embedder to report the size
of external string data for inclusion in DevTools heap snapshots. This
change implements the new virtual functions so that heap snapshots can
become more accurate, particularly if a string is parked.

This change also corrects an inaccuracy in
ParkableStringImpl::MemoryFootprintForDump and
ParkableStringManager::ComputeStatistics, which both failed to account
for the size of the StringImpl when the string is unparked.

Bug: 40734351, 41448424
Change-Id: Ia8ff64753137df3bf65de5fb7419611c40764e07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6221187
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1420026}
NOKEYCHECK=True
GitOrigin-RevId: c8d73dcf2c6793a6af79b4b874f37bcc2fc6ac78