Make UpSampler::Reset() call member Reset() instead of dropping memory
Currently UpSampler::Reset() will reset the direct_convolver_ and
simple_fft_convolver_ unique pointers, which will result in an access
violation if Process() is called after calling Reset().

Instead, call the corresponding Reset() functions as was probably
intended.

This method is currently only used by the WaveShaper Reset() function,
which is currently unused.  That is why it has not caused problems yet.
I plan to use this function in the future though, so we should fix it.

Change-Id: Ia28663f8dd066a97a416bb97ce7b18790aea31eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6265545
Commit-Queue: Michael Wilson <mjwilson@chromium.org>
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1420139}
NOKEYCHECK=True
GitOrigin-RevId: 7dece07b1e40586f4b88751c5d93607cece2ac28