[A11y] Address crash when accessing focused AXObject and updating tree
The crash occurs because while adding children back to the container of
the focused textfield, the textfield is still temporarily parentless.
While that is true, using FocusedObject() is a problem because it calls
AXObject::IsIgnored() which for the focused input's object, which
wants to update cached values, and that asserts there must be a parent.

Added a multiselect check as well, since multiselect widgets don't
need to run this code. It doesn't affect the crash, but it's a
simple check and makes things more efficient in that case.

Fixed: 396048653,395595617
Bug: 393260687,380882534,394053341
Change-Id: I8d4beed135f2e849680480b3cc3431e00349bad3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6264886
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Lucas Radaelli <lucasradaelli@google.com>
Cr-Commit-Position: refs/heads/main@{#1420217}
NOKEYCHECK=True
GitOrigin-RevId: fd9cbe1d035b62cf68789bf1bf7e469b0cc9e432