Measure usage of Geolocation PositionOptions enableHighAccuracy
This will allow us to know how many Geolocation requests come with the
enableHighAccuracy flag set to true. Note that we already collect use
counters for all calls to getCurrentPosition and watchPosition, so it
should be relatively easy to compare the two.

Change-Id: I75d57291c2ad0881effa31cf4db9b38a185304fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6269656
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1420466}
NOKEYCHECK=True
GitOrigin-RevId: dc5855d625afce1454ca2bb91b560a52479815a7