Remove testContent setter optimization when mutation observers present
See linked bugs for more context. But this optimization is observable
(and non-interoperable) when mutation observers are present. This CL
removes that optimization in the case that there are mutation observers
present on the document.

This is tested via the MutationObserver-textContent.html WPT.

Fixed: 41095015, 40553863, 391394132
Change-Id: I9fd1e4e7d5179388c01fa873fb1116e7f10c0949
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6255192
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1420676}
NOKEYCHECK=True
GitOrigin-RevId: 9be8c38c3b437f8e6fb63846c5fdfcef399e9a48