Resolve css scale interpolation number before interpolation
This is done to fix the case when we have a neutral frame, which will
end up being a number type and some percentage frame, but we are not
allowed to sum number and percentage.
So the fix is to resolve percentage to number before interpolation.
It can give incorrect result if values
have some relative units inside that are also being interpolated.

We would be able to do it properly by dividing the percentage part
by 100% during interpolation, once we can divide units properly,
so leaving TODO for now.

Fixed: 396584141
Change-Id: Iac9e16047237264e3fd56b5dfb9e4f0f2684c6c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6276493
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Daniil Sakhapov <sakhapov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1421095}
NOKEYCHECK=True
GitOrigin-RevId: aa5ee3e7654f683abf243a0b91cdda43d709dc98