Disallow prefixed properties in highlight pseudos
The spec was updated to disallow prefixed properties in CSS Highlight
Pseudo element. Specifically:

"Vendor-prefixed properties such as -webkit-text-fill-color are not
applicable to the highlight pseudo-elements."

Fix this and update the tests that were testing for support to test for
non-support.

https://drafts.csswg.org/css-pseudo-4/#highlight-styling

Fixed: 390229581
Change-Id: Ieeb51411987ba592ea62a3972cff02b3ca8f2e43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6277924
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1421370}
NOKEYCHECK=True
GitOrigin-RevId: c0159ab76c22769a155cb952bf2d6b44afda7b77