[functions] Handle {}-wrapper at parse-time
Each argument to a custom function is a "comma containing production",
which requires some special handling of "{}" [1]. In short, commas
and "{}" blocks are not allowed in arguments, unless the whole argument
is wrapped in "{}". This makes it possible to pass values containing
a comma as a single argument, e.g.: --foo({1, 2}, 3).

This CL just implements the parse-time enforcement of these rules;
a subsequent CL will make StyleCascade::ResolveFunctionInto aware
of this behavior.

[1] https://drafts.csswg.org/css-values-5/#component-function-commas

Bug: 325504770, 390205875
Change-Id: Ic9abeb9884b503e31ae3b5ffa0b5ff58c74d5677
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6276539
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1421638}
NOKEYCHECK=True
GitOrigin-RevId: 5d45c2c3b6ddbc28a168dcd2f1afd97331ac800b