[libaom] Remove libaom_buildflags.
libaom license was included in the LICENSE file generated by the
licenses.py tool even when enable_libaom was set to false. This was
caused by the libaom_buildflags target - it was in the
third_party/libaom directory and so it was presumed to be covered by the
corresponding license. OTOH the target is not necessary since the build
flags is included in media_buildflags.

Bug: 396724495
Change-Id: Ic6bb8dde199123360dbb246e79280f431a6525f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6277636
Auto-Submit: Sergey Ulanov <sergeyu@chromium.org>
Commit-Queue: Eugene Zemtsov <eugene@chromium.org>
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1421717}
NOKEYCHECK=True
GitOrigin-RevId: 0247535ca44c9640e98ed1e233aa2ac7be98e402