[Signature-based SRI] Simplify path-based tests.
Currently, our tentative WPTs rely on a resource file that lives in the
`/tentative` directory. This directory impacts the signature in some
cases (e.g. the `@path` derived component), and it would be ideal to
avoid that as we introduce more cases where it might occur. This CL
moves the resouce file out of the `/tentative` directory so that the
signatures will remain stable even as the tests shift from tentative to
non-tentative.

Bug: 383409584
Change-Id: Ie8cbba80e856e4a1b051e80e095875cefe7c5aa6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6268259
Commit-Queue: Mike West <mkwst@chromium.org>
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1421795}
NOKEYCHECK=True
GitOrigin-RevId: cb05cc9ddb5f9fbcdcf6ddc8367a4bfc99aa2d8b