Consider the 'preserveAspectRatio' attribute on <feImage>
An overriding preserveAspectRatio (pAR) value from the element was not
applied when laying out or painting the referenced image - it was only
considered when computing the destination/source rects.

Provide the pAR value in the form of a view spec to the image wrapper.

Fixed: 397258346
Change-Id: I5624762be7fbe89f2192c0ce7f0e54af28fd7740
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6276495
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1421824}
NOKEYCHECK=True
GitOrigin-RevId: 8692b015aaf0acdcd12ff3cea587ddd99d8975ea