[Blink] Make StaticBitmapImage::GetSkImageInfo() private
This CL converts the last remaining public callsite of
StaticBitmapImage::GetSkImageInfo() to instead construct an
SkImageInfo from the getters for the individual fields and moves the
method to private. We will follow up by changing the getters to
themselves be virtual and eliminating GetSkImageInfo() altogether.

Bug: 371227617
Change-Id: Id56965b1c55f23d2ab76501eafaad0c4377e4e3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6276095
Reviewed-by: Saifuddin Hitawala <hitawala@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1421837}
NOKEYCHECK=True
GitOrigin-RevId: 78428c10933b95dabd6308783b3b0da3e572af7d