[CSS][SVG] Fix inheritance for all :visited styles
This CL fixes the failure in the WPT color-inherit-link-visited.svg
[1]

The WebKit commit [2]
introduces logic and tests to ensure that elements inside a link
inherit a computed color from another element.
This change was intended to cover both CSS and SVG elements.
However, in Chromium, the WPT for SVG started failing.

The issue occurred because we only expand visited properties when
inside a visited link. Consequently, the visited properties did not
inherit values from their unvisited counterparts. The WPT failure is
related to the `internal-visited-fill` property, but similar issues
can arise with other CSS properties like
`internal-visited-text-emphasis-color`.

This CL resolves the WPT failure and the broader problem of visited
properties not inheriting unvisited values. In this CL we apply the
unvisited values to the visited properties during the initial style
calculation of the visited link element.

This CL also includes new WPTs for all the inherited visited
properties for which we are applying unvisited values before
`ExpandCascade`.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/external/wpt/svg/styling/color-inherit-link-visited.svg?q=color-inherit-link-visited&ss=chromium%2Fchromium%2Fsrc

[2] https://github.com/WebKit/WebKit/commit/f93d71021b6613c26c7db8da6be03f88169d32c5

Note: Even with this change there will be a bug corresponding to
explicit inheritance (of non-inherited visited properties).

Bug: 388400236, 40356084
Change-Id: I36dd328e6b9435e3baf1baab71867f8de7efd263
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6137310
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Divyansh Mangal <dmangal@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1421874}
NOKEYCHECK=True
GitOrigin-RevId: b39d6582c8e5f81c3308aa7985a92c3475dc2c36