SmartCard: Add watcher for tracking smart card connections/usage
Context: go/web-smart-card-usage-tracking

This introduces a new Mojo interface SmartCardConnectionWatcher and adds
it as an optional argument for the Connect(). When it's passed to the
platform-specific PC/SC implementation, it'll be used to inform:
- Platform implementation - of the grant expiry
- Browser - about the card usage, enabling to show indicator and impose
  user attention requirements

For now this does nothing except incrementing timestamps and setting
when there's any active connection - this CL is separate due to its
size. Future work:
- Add grant expiry based on this
- Show indicator when grants are in use

Bug: 389631854
Change-Id: Ia8f226d514ba973f56231eb683aa76e282c98560
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6213591
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Andrew Rayskiy <greengrape@google.com>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Simon Hangl <simonha@google.com>
Commit-Queue: Luke Klimek <zgroza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1421897}
NOKEYCHECK=True
GitOrigin-RevId: 288960ba66da3d15f1bfd392b783743538406990