corner-shape: support overflow clipping
Update the PaintPropertyTree as well as the layer chunks
to use the path-based rounded rect instead of the SKRect.

This makes it so that clipping a corner-shaped box goes
through the same code-path as rendering clip-path.

Note that currently only bevel rendering is enabled, however this
is going to be valid for the other curvatures that are drawn as
a rounded rect path.

Bug: 394900356
Change-Id: I10f393d4bfce1cea6844bc562223856240c48658
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6275072
Commit-Queue: Noam Rosenthal <nrosenthal@chromium.org>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1421940}
NOKEYCHECK=True
GitOrigin-RevId: 24f7d3576211615e00bbcdda13d5f8dc4e01ce59