Remove the 'variants' param from instantiated variants
The 'variants' parameter is used to generate the list of Variant
instances for a given test. It's not useful to keep it around once the
Variant objects are created, or else, each Variant object ends up
carrying the full list of all variant configs in it's params.

Bug: 40207206
Change-Id: I462ad791836dd439385b8ad7a5afd9c278343279
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6204125
Reviewed-by: Andres Ricardo Perez <andresrperez@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422138}
NOKEYCHECK=True
GitOrigin-RevId: 329d98edac406422c4d4d09ba021a1076a97f880