Fix paint invalidation for ::selection style changes
The existing code for invalidating paint on a ::selection
style change assumed the old model of inheritance for ::selection,
where the style only applied to the element it matched for. Under
that assumption it only checked the direct text children of the
element.

With Highlight Inheritance the assumption is invalid, and the
entire subtree may be affected by the style change. So process
the subtree checking for selected content.

Fixed: 395934180
Change-Id: Ifa8ac3309d79ce4b3696ed042cb822361a7b490c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6283486
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422232}
NOKEYCHECK=True
GitOrigin-RevId: 7a6ed3a2c0f5632a2dd5a2212fcb7b33f29a2c9c