Feature LimitLayerMergeDistance
The feature has a parameter as the upper limit of the distance between
PendingLayers that can be merged. This reduces the cost of the
layerization algorithm. Especially, this avoids the worst complexity
scenario where there are many PendingLayers and they don't overlap but
can't be merged.

Bug: 40929618
Change-Id: I5f41f3e387f384eca423271a5b80af36e82045b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6281410
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422279}
NOKEYCHECK=True
GitOrigin-RevId: 3fbdbcc6108fc85ce394288ad813036233ac9c79