GetSourceImageForCanvas: Do not DCHECK AlphaDisposition
The two options for AlphaDisposition are kPremultiplyAlpha and
kDontChangeAlpha.

There used to be an option to un-multiply alpha. That is no longer
there, and so kPremultiplyAlpha is more restrictive than the other
option, and so it doesn't make sense to DCHECK this anymore.

Follow-up changes will remove the parameter.

Bug: 40277213
Change-Id: I68eac72d875cc18e632cdc7628d8177aa4ba04ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6281585
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422474}
NOKEYCHECK=True
GitOrigin-RevId: c8f3998bad48506e3698844c4561da85fd5cba19