Nested VT: Introduce a for-subframe-snapshot bool to VT resource id.
This patch adds a for-subframe-snapshot bool to VT resource id and
uses it to disable CaptureAndView mode for view transition tokens that
include subframe captures. The reason for this is that subframe capture
is used to hold iframe's screenshot in place while the underlying
content is updated. It is by design that it doesn't follow
CaptureAndView method.

There may be follow up work in referenced bug to add tests to verify
that groups in iframes with this mode behave correctly.

This patch 4 out of 5 to implement the feature.

R=pdr@chromium.org

Bug: 358119267
Change-Id: I14664362cf853d50b275f43a30832e1b2bdd213e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6259475
Reviewed-by: Joe Mason <joenotcharles@google.com>
Commit-Queue: Vladimir Levin <vmpstr@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422643}
NOKEYCHECK=True
GitOrigin-RevId: f66a7b9a1e6f7a3d751302cdb3866658489b718b