[SPC] Fix error thrown for activationless create in cross-origin iframe
Calling navigator.credentials.create() in a cross-origin iframe without
a user activation should result in a NotAllowedError, per the [WebAuthn
spec]. However for 'payment' credentials, the SPC spec used to say to
throw a SecurityError instead, and Chrome's implementation did exactly
that.

This contradiction was fixed in the SPC spec in
https://github.com/w3c/secure-payment-confirmation/pull/281, so that
NotAllowedError is always the correct error to throw. This CL follows
through in the Chrome implementation to align payment credential create
with non-payment credentials.

Because this change is a breaking one (the type of the error changes),
the fix is behind a flag for now. We will be connecting with partners
and following a standard web platform process for handling breaking
changes.

[WebAuthn spec]: https://w3c.github.io/webauthn/#sctn-createCredential

Bug: 41484826
Change-Id: I133017586cac89b94eec1738831356d0398de0e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6284564
Reviewed-by: Florin Malita <fmalita@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Auto-Submit: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Florin Malita <fmalita@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422745}
NOKEYCHECK=True
GitOrigin-RevId: fb08fd95b53e39978da603ae080ac745f5deeac9